Add screen navigation logic
Tab Bar: can only click on complete or in progress screens. can only move on to the next screen when previous is filled out
Screens: created "Next" button which validates inputs and goes to the next screen when satisfied.
WIP: remove screen swiping capability
WIP: complete logic for Location page and beyond.
WIP: grey out tabs which are not able to be accessed yet